refactor: move device query to productlist model
