--- frontend/lib/algolia/server/index.ts
@@ -1,1 +1,0 @@
-export * from './findDescendantDevicesWithProducts';
